# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ADOCAO API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/animal/{idUsuario}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Mostrar usuario",
                "description": "Consultar informações do usuario",
                "operationId": "mostrarUsuario",
                "parameters": [
                    {
                        "name": "idUsuario",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "usuario": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Matheus Eduardo"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "contato.matheused@gmail.com"
                                                },
                                                "foto": {
                                                    "type": "string",
                                                    "example": "null"
                                                },
                                                "endereco": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": "1"
                                                        },
                                                        "rua": {
                                                            "type": "string",
                                                            "example": "João dos Santos"
                                                        },
                                                        "numero": {
                                                            "type": "integer",
                                                            "example": "231"
                                                        },
                                                        "complemento": {
                                                            "type": "string",
                                                            "example": "null"
                                                        },
                                                        "bairro": {
                                                            "type": "string",
                                                            "example": "Ipê"
                                                        },
                                                        "cidade": {
                                                            "type": "string",
                                                            "example": "Poços de Caldas"
                                                        },
                                                        "estado": {
                                                            "type": "string",
                                                            "example": "Minas Gerais"
                                                        },
                                                        "cep": {
                                                            "type": "string",
                                                            "example": "37704205"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Não autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Animal não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/usuario": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Listagem de usuarios",
                "description": "Lista com paginação de todos os usuarios.",
                "operationId": "listarUsuarios",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Buscar por numero de paginação"
                    },
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "Buscar pelo nome do usuário"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successo!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Matheus Eduardo"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "contato.matheused@gmail.com"
                                        },
                                        "foto": {
                                            "type": "string",
                                            "example": "null"
                                        },
                                        "endereco": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "rua": {
                                                    "type": "string",
                                                    "example": "João dos Santos"
                                                },
                                                "numero": {
                                                    "type": "integer",
                                                    "example": "231"
                                                },
                                                "complemento": {
                                                    "type": "string",
                                                    "example": "null"
                                                },
                                                "bairro": {
                                                    "type": "string",
                                                    "example": "Ipê"
                                                },
                                                "cidade": {
                                                    "type": "string",
                                                    "example": "Poços de Caldas"
                                                },
                                                "estado": {
                                                    "type": "string",
                                                    "example": "Minas Gerais"
                                                },
                                                "cep": {
                                                    "type": "string",
                                                    "example": "37704205"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}